Implemted: The code works great! The bird works good and the game acually works
Addressed issue: The score was giving the player 1 even without start the game

Things for the future:
Add a more dyanamic touch for the pipes they have no sensitivity for the bird. I can play with it in the future and have a more dynamic idea. I can add more fun game like and challeging.